+++ CHANGELOG.md
@@ -23,6 +23,10 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/virtual-users/compare/release-3.11.1...master
 
+### Added
+- Add `VirtualUserNodeResult` and `VirtualUserResult` to formalize the filesystem result format.
+- Write results to the `VirtualUserBehavior.results` path.
+
 ## [3.11.1] - 2020-04-22
 [3.11.1]: https://github.com/atlassian/virtual-users/compare/release-3.11.0...release-3.11.1